update some work scripts

+++ home-manager/profiles/dell5560/scripts/cvm.sh
@@ -1,13 +1,45 @@
 #!/usr/bin/env bash
+set -e -o pipefail
 
-LOCAL_ROOT="$HOME/indeed"
-REMOTE_HOST="$USER.cvm.indeed.net"
-REMOTE_USER="$USER"
-REMOTE_ROOT="/home/$USER/indeed"
+DEFAULT_LOCAL_ROOT="$HOME/indeed"
+DEFAULT_REMOTE_ROOT="/home/$USER/indeed"
+DEFAULT_REMOTE_USER="$USER"
+DEFAULT_REMOTE_HOST="$USER.cvm.indeed.net"
+CONFIG_PATH="$HOME/.cvm"
+
+bold=$(tput bold)
+normal=$(tput sgr0)
 
 function cvm() (
+
+  function cvm_init() {
+    local changed=false
+    [ -z "$LOCAL_ROOT" ] && read -r -p "${bold}Path to indeed folder on your laptop [$DEFAULT_LOCAL_ROOT]:${normal} " && changed=true
+    [ -z "$LOCAL_ROOT" ] && LOCAL_ROOT=$DEFAULT_LOCAL_ROOT
+
+    [ -z "$REMOTE_HOST" ] && read -r -p "${bold}Your cvm hostname [$DEFAULT_REMOTE_HOST]: ${normal}" && changed=true
+    [ -z "$REMOTE_HOST" ] && REMOTE_HOST=$DEFAULT_REMOTE_HOST
+
+    [ -z "$REMOTE_USER" ] && read -r -p "${bold}Your user name on $REMOTE_HOST [$DEFAULT_REMOTE_USER]: ${normal}" && changed=true
+    [ -z "$REMOTE_USER" ] && REMOTE_USER=$DEFAULT_REMOTE_USER
+
+    [ -z "$REMOTE_ROOT" ] && read -r -p "${bold}Path to indeed folder on $REMOTE_HOST [$DEFAULT_REMOTE_ROOT]: ${normal}" && changed=true
+    [ -z "$REMOTE_ROOT" ] && REMOTE_ROOT=$DEFAULT_REMOTE_ROOT
+
+    if $changed
+      then cat <<CONFIG > "$CONFIG_PATH"
+LOCAL_ROOT=$LOCAL_ROOT
+REMOTE_HOST=$REMOTE_HOST
+REMOTE_USER=$REMOTE_USER
+REMOTE_ROOT=$REMOTE_ROOT
+CONFIG
+    fi
+  }
+
 	function cvm_sync() {
+    [[ ! "$PWD" =~ "$LOCAL_ROOT/".* ]] && return 1
 		REMOTE_PATH=$(echo "$PWD" | sed "s#$LOCAL_ROOT#$REMOTE_ROOT#g")
+    echo "Synchronizing $PWD to $REMOTE_USER@$REMOTE_HOST:$REMOTE_PATH"
 		rsync -aP --delete --exclude build/ --exclude node_modules/ --exclude .gradle/ "$PWD/." $USER@$REMOTE_HOST:$REMOTE_PATH
 	}
 
@@ -15,15 +47,29 @@ function cvm() (
 		ssh "$REMOTE_HOST"
 	}
 
-	function cvm_exec() {
+	function cvm_exec_in_synced_folder() {
+    [[ ! "$PWD" =~ "$LOCAL_ROOT/".* ]] && return 1
 		REMOTE_PATH=$(pwd | sed "s#$LOCAL_ROOT#$REMOTE_ROOT#g")
-		# Run sync only if we are in a subfolder of $REMOTE_ROOT.
-		[ $? -eq 0 ] && [ "$REMOTE_PATH" != "$REMOTE_ROOT" ] && cvm_sync
-		ssh ${REMOTE_USER}@${REMOTE_HOST} "cd $REMOTE_PATH && echo $* | bash --login"
+
+		cvm_sync
+
+    echo "Executing '$*' at $REMOTE_USER@$REMOTE_HOST:$REMOTE_PATH"
+    ssh ${REMOTE_USER}@${REMOTE_HOST} "cd $REMOTE_PATH && echo $* | bash --login"
 	}
 
-	[ $# -eq 0 ] && cvm_connect
-	[ $# -gt 0 ] && cvm_exec "$@"
+	function cvm_exec() {
+    echo "Executing '$*' at $REMOTE_USER@$REMOTE_HOST:~"
+    ssh ${REMOTE_USER}@${REMOTE_HOST} "echo $* | bash --login"
+	}
+
+  [ -f "$CONFIG_PATH" ] && source "$CONFIG_PATH"
+  cvm_init
+  if [ $# -eq 0 ];
+    then cvm_connect
+  elif [[ "$PWD" =~ "$LOCAL_ROOT/".* ]]
+    then cvm_exec_in_synced_folder "$@"
+    else cvm_exec "$@"
+  fi
 )
 
 cvm "$@"
